Xion and Xehanort Data Portals Force to Yozora
With this there were changes in the cutscene that plays in Radiant Garden after defeating Xion and Xehanort where the game sends you to Yozora. (cutscene_rg01_rg074). This was made to take the player to "rg_redirector" so that the player would teleport back into the Garden of Assemblage with the real data portals instead of going to Yozora. While their portals remain blue after defeating them this is purely cosmetic. Fixes #6