All QoL changes and finals changes for 1.1
This addresses all QoL additions being added to both the website and the game itself as well as any remaining bugs that have been found since release. Any other QoL additions will be added in 1.2 since the difference between 1.0 and 1.1 is now so large it is finally time for an update. Fixes #31 #57 #32 #62 #59 #61 #58 #66 #67